Data Collection Ready.

+++ PA2/data_collection/multi_level_feedback_queue.c
@@ -42,16 +42,24 @@ void myfunction(int param){
 	}
 }
 /************************************************************************************************/
+
 struct process_description{
 	int workload_time;
 	pid_t pid;
 	int id;
-	struct timespec end_time, start_time;
+	struct timespec end_time;
+	double processing_time;
 };
-int main(int argc, char const *argv[])
-{
+
+int main(int argc, char const *argv[]) {
+
+	if(argc < 2){
+		printf("USAGE: ./a.out QUANTUM\n");
+		exit(1);
+	}
+	int QUANTUM = atoi(argv[1]);
+
 	pid_t pid1, pid2, pid3, pid4;
-	int running1, running2, running3, running4;
 
 	// Create all four processes
 	pid1 = fork();
@@ -85,30 +93,29 @@ int main(int argc, char const *argv[])
 	/************************************************************************************************ 
 		At this point, all  newly-created child processes are stopped, and ready for scheduling.
 	*************************************************************************************************/
-
 	// Store process information in a struct
 	struct process_description processes[4], temp_process;
-	for(int i = 0; i < 4; i++){processes[i].end_time.tv_nsec = 0;}
+	for(int i = 0; i < 4; i++){
+		processes[i].end_time.tv_nsec = 0; 
+		processes[i].processing_time = 0;
+		processes[0].id = i+1;
+	}
 	processes[0].workload_time = WORKLOAD1;
 	processes[0].pid = pid1;
-	processes[0].id = 1;
 
 	processes[1].workload_time = WORKLOAD2;
 	processes[1].pid = pid2;
-	processes[1].id = 2;
 
 	processes[2].workload_time = WORKLOAD3;
 	processes[2].pid = pid3;
-	processes[2].id = 3;
 
 	processes[3].workload_time = WORKLOAD4;
 	processes[3].pid = pid4;
-	processes[3].id = 4;
 
 	// Start timing
-    struct timespec start, end, temp1, proc_start;
-	// Start Overhead calculation
+    struct timespec start, end, proc_end[4], temp1, temp2, proc_start;
     clock_gettime(CLOCK_MONOTONIC_RAW, &start);
+    clock_gettime(CLOCK_MONOTONIC_RAW, &proc_start);
 
 
 	/************************************************************************************************
@@ -118,42 +125,85 @@ int main(int argc, char const *argv[])
 		to be implemented.
 	************************************************************************************************/
 
-	// Find Scheduling Order (it is the same order as it comes in)
+	int running[4] = {1,1,1,1};
+
+	clock_gettime(CLOCK_MONOTONIC_RAW, &temp1);
+	kill(pid1, SIGCONT);
+	usleep(QUANTUM);
+	kill(pid1, SIGSTOP);
+	clock_gettime(CLOCK_MONOTONIC_RAW, &temp2);
+	processes[0].processing_time += (temp2.tv_sec - temp1.tv_sec) + (double)(temp2.tv_nsec - temp1.tv_nsec) / 1000000000;
+	
+	clock_gettime(CLOCK_MONOTONIC_RAW, &temp1);
+	kill(pid2, SIGCONT);
+	usleep(QUANTUM);
+	kill(pid2, SIGSTOP);
+	clock_gettime(CLOCK_MONOTONIC_RAW, &temp2);
+	processes[1].processing_time += (temp2.tv_sec - temp1.tv_sec) + (double)(temp2.tv_nsec - temp1.tv_nsec) / 1000000000;
+
+	clock_gettime(CLOCK_MONOTONIC_RAW, &temp1);
+	kill(pid3, SIGCONT);
+	usleep(QUANTUM);
+	kill(pid3, SIGSTOP);
+	clock_gettime(CLOCK_MONOTONIC_RAW, &temp2);
+	processes[2].processing_time += (temp2.tv_sec - temp1.tv_sec) + (double)(temp2.tv_nsec - temp1.tv_nsec) / 1000000000;
+
+	clock_gettime(CLOCK_MONOTONIC_RAW, &temp1);
+	kill(pid4, SIGCONT);
+	usleep(QUANTUM);
+	kill(pid4, SIGSTOP);
+	clock_gettime(CLOCK_MONOTONIC_RAW, &temp2);
+	processes[3].processing_time += (temp2.tv_sec - temp1.tv_sec) + (double)(temp2.tv_nsec - temp1.tv_nsec) / 1000000000;
+
+	for(int i = 0; i < 4; i++){
+		waitpid(processes[i].pid, &running[i], WNOHANG);
+		if (running[i] == 0 && processes[i].end_time.tv_nsec == 0){
+			clock_gettime(CLOCK_MONOTONIC_RAW, &processes[i].end_time);
+		}
+
+	}
 
 	int status;
-    clock_gettime(CLOCK_MONOTONIC_RAW, &proc_start);
 	for(int i = 0; i < 4; i++){
-		clock_gettime(CLOCK_MONOTONIC_RAW, &processes[i].start_time);
-		kill(processes[i].pid, SIGCONT);
-		waitpid(processes[i].pid, &status, 0);
-		clock_gettime(CLOCK_MONOTONIC_RAW, &processes[i].end_time);
+		if(running[i]){
+			clock_gettime(CLOCK_MONOTONIC_RAW, &temp1);
+			kill(processes[i].pid, SIGCONT);
+			waitpid(processes[i].pid, &status, 0);
+			clock_gettime(CLOCK_MONOTONIC_RAW, &processes[i].end_time);
+			processes[i].processing_time += (processes[i].end_time.tv_sec - temp1.tv_sec) + (double)(processes[i].end_time.tv_nsec - temp1.tv_nsec) / 1000000000;
+
+		}
 	}
+
 	/************************************************************************************************
 		- Scheduling code ends here
 	************************************************************************************************/
 	
 	// Record End time
     clock_gettime(CLOCK_MONOTONIC_RAW, &end);
-
-	double turnaround_time[4], processing_time[4];
+	
+	double turnaround_time[4];
 	for(int i = 0; i < 4; i++){
 		turnaround_time[i] = (processes[i].end_time.tv_sec - start.tv_sec) + (double)(processes[i].end_time.tv_nsec - start.tv_nsec) / 1000000000;
-		processing_time[i] = (processes[i].end_time.tv_sec - processes[i].start_time.tv_sec) + (double)(processes[i].end_time.tv_nsec - processes[i].start_time.tv_nsec) / 1000000000;
 	}
 
 	double total_time = (end.tv_sec - start.tv_sec) + (double)(end.tv_nsec - start.tv_nsec) / 1000000000;
 	double scheduling_time = (proc_start.tv_sec - start.tv_sec) + (double)(proc_start.tv_nsec - start.tv_nsec) / 1000000000;
-	double context_switch_time = total_time - scheduling_time - processing_time[0] - processing_time[1] - processing_time[2] - processing_time[3];
+	double context_switch_time = (total_time - scheduling_time
+		- processes[0].processing_time - processes[1].processing_time 
+		- processes[2].processing_time - processes[3].processing_time);
 	
 	// Print Results
-    // printf("Turnaround Times: {%0.5f, %0.5f, %0.5f, %0.5f}.\n", turnaround_time[0], turnaround_time[1], turnaround_time[2], turnaround_time[3]);
-    // printf("Processing Times: {%0.5f, %0.5f, %0.5f, %0.5f}.\n", processing_time[0], processing_time[1], processing_time[2], processing_time[3]);
+    // printf("Turnaround Times: {%0.5f, %0.5f, %0.5f, %0.5f}.\n", turnaround_time[0], 
+	// 	turnaround_time[1], turnaround_time[2], turnaround_time[3]);
+    // printf("Processing Times: {%0.5f, %0.5f, %0.5f, %0.5f}.\n", processes[0].processing_time, 
+	// 	processes[1].processing_time, processes[2].processing_time, processes[3].processing_time);
 	// printf("Context Switch Time: %0.8f\n", context_switch_time);
 	// printf("This operation took %0.5f seconds.\n", total_time);
-
+	
 	double avg_resp_time = (turnaround_time[0] + turnaround_time[1] + turnaround_time[2] + turnaround_time[3]) / 4; 
 	// printf("Average Response Time: %0.8f\n", avg_resp_time);
-	printf("%d, %lf, %lf\n", 0, avg_resp_time, context_switch_time);
+	printf("%d, %0.8f, %0.8f\n", QUANTUM, avg_resp_time, context_switch_time);
 
 	return 0;
 }
